+++ core/src/main/towers/RiotShield.java
@@ -15,7 +15,7 @@ public class RiotShield extends Building implements Voter {
 
     private static final String NAME = "RIOT SHIELD";
     private static final int COST = 50;
-    private static final int COOLDOWN = 100;
+    private static final float COOLDOWN = 100;
     private Party party = PartyFactory.Democrat(6); // Just for now
     private static String DESCRIPTION = "Pushes back nearby enemies";
     private static final int RADIUS = 500;
